XIP: Add custom bootx command to boot XIP kernel
Can use bootm anymore because we are not booting a uImage. Also just can't
use 'go' because we need to modify R0,R1,R2 to pass arguments to the
kernel.

Note that bootargs is not used when bootx is used, so make sure your
command line in your DT blob is what you want.

Also make sure your DT blob is in RAM before using this command because
the kernel will not like accessing a DT pointer to the SPI Flash area.